Refactored fake user generation. Moved to helper file
